Change `quota` type to `unit64_t` to avoid macOS compilation issue

+++ fdbclient/ManagementAPI.actor.cpp
@@ -2460,17 +2460,17 @@ bool schemaMatch(json_spirit::mValue const& schemaValue,
 	}
 }
 
-void setStorageQuota(Transaction& tr, StringRef tenantName, size_t quota) {
+void setStorageQuota(Transaction& tr, StringRef tenantName, uint64_t quota) {
 	auto key = storageQuotaKey(tenantName);
-	tr.set(key, BinaryWriter::toValue<size_t>(quota, Unversioned()));
+	tr.set(key, BinaryWriter::toValue<uint64_t>(quota, Unversioned()));
 }
 
-ACTOR Future<Optional<size_t>> getStorageQuota(Transaction* tr, StringRef tenantName) {
+ACTOR Future<Optional<uint64_t>> getStorageQuota(Transaction* tr, StringRef tenantName) {
 	state Optional<Value> v = wait(tr->get(storageQuotaKey(tenantName)));
 	if (!v.present()) {
-		return Optional<size_t>();
+		return Optional<uint64_t>();
 	}
-	return BinaryReader::fromStringRef<size_t>(v.get(), Unversioned());
+	return BinaryReader::fromStringRef<uint64_t>(v.get(), Unversioned());
 }
 
 std::string ManagementAPI::generateErrorMessage(const CoordinatorsResult& res) {
